% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_function}
\alias{example_function}
\title{Example Function}
\usage{
example_function(x)
}
\arguments{
\item{x}{An input parameter.}
}
\value{
What gets returned by the function?
}
\description{
Example paragraph describing
what the function does.
}
\examples{
example_function(5)
}
